requirements.txt
fastapi
uvicorn[standard]
duckdb
geopandas
shapely
python-dotenv
overturemaps
APScheduler